WIP Don't save window state
There must be people who only use a single terminal window, and it is
always in the same position and size.  Or maybe maximized (I actually
know such a person).  But that only makes sense for small laptop
screens.  I am not such a person, though.  Even on a laptop I rarely
maximize terminal windows.  And if I do, it's only vertically (there
does not seem to be support for this).  And even if I maximize one
window, I don't want the next window I open to be in the same spot and
also maximized.  That's why there are shortcut keys to maximize
windows horizontally and vertically.

To complete this we have to add a setting to control window state
saving.

Or if we maintain our own fork just remove the state management
altogether (including from schema and stuff).